+++ prog/export_test.go
@@ -32,7 +32,7 @@ func initTargetTest(t *testing.T, os, arch string) *Target {
 }
 
 func randSource(t *testing.T) rand.Source {
-	seed := int64(time.Now().UnixNano())
+	seed := time.Now().UnixNano()
 	if os.Getenv("TRAVIS") != "" {
 		seed = 0 // required for deterministic coverage reports
 	}
